# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AUZEF Chat Analiz API",
    "description": "AUZEF chatbot mesajlarından sık sorulan soru ve tema analizi.\n\nTüm hata cevapları RFC 9457 Problem Details biçimindedir ve her zaman `type`, `title`, `status`, `code`, `detail`, `trace_id` alanlarını taşır. `retry_after` YALNIZCA 429 cevaplarında bulunur ve başka hiçbir cevapta `null` olarak dahi yer almaz.\n\nTüm tarihler UTC ISO 8601'dir ve `Z` ile biter (`YYYY-MM-DDTHH:MM:SS.sssZ`).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Dosya yükle ve profil işini kuyruğa al",
        "operationId": "create_upload_api_v1_uploads_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Opsiyonel. Aynı anahtar + aynı gövde ile tekrarlanan istek, ilk isteğin 202 cevabını aynen döndürür (replay). Aynı anahtar + FARKLI gövde ise 409 `JOB_CONFLICT` üretir. Kayıt 24 saat saklanır.\n\nReplay, orijinal 202'yi döndürür ve istekle gelen yeni `X-OpenRouter-Key` header'ını YOK SAYAR: anahtar orijinal job'a bağlıdır. Aksi hâlde anahtar rotasyonundan sonraki bir retry, ölmüş bir job'ı sessizce diriltirdi.\n\nFingerprint ve saklama anahtarı ADR-0002 #3'te byte seviyesinde tanımlıdır; secret/header değerleri fingerprint'e girmez.",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "example": "8f14e45f-ceea-467a-9f6b-2c1d3e4a5b6c"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_api_v1_uploads_post"
              },
              "examples": {
                "XLSX dosyası": {
                  "summary": "file alanında binary .xlsx içeriği",
                  "value": {
                    "file": "(binary .xlsx content)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCreated"
                },
                "examples": {
                  "Kuyruğa alındı": {
                    "value": {
                      "upload_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                      "status": "queued"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "409": {
            "description": "RFC 9457 Problem Details — olası kodlar: JOB_CONFLICT",
            "x-error-codes": [
              "JOB_CONFLICT"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "JOB_CONFLICT": {
                    "value": {
                      "type": "/errors/job-conflict",
                      "title": "İşlem durumu bu isteğe uygun değil",
                      "status": 409,
                      "code": "JOB_CONFLICT",
                      "detail": "İşlem durumu bu isteğe uygun değil (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "413": {
            "description": "RFC 9457 Problem Details — olası kodlar: UPLOAD_TOO_LARGE",
            "x-error-codes": [
              "UPLOAD_TOO_LARGE"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "UPLOAD_TOO_LARGE": {
                    "value": {
                      "type": "/errors/upload-too-large",
                      "title": "Dosya boyutu sınırı aşıldı",
                      "status": 413,
                      "code": "UPLOAD_TOO_LARGE",
                      "detail": "Dosya boyutu sınırı aşıldı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "415": {
            "description": "RFC 9457 Problem Details — olası kodlar: UPLOAD_INVALID_TYPE",
            "x-error-codes": [
              "UPLOAD_INVALID_TYPE"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "UPLOAD_INVALID_TYPE": {
                    "value": {
                      "type": "/errors/upload-invalid-type",
                      "title": "Desteklenmeyen dosya türü",
                      "status": 415,
                      "code": "UPLOAD_INVALID_TYPE",
                      "detail": "Desteklenmeyen dosya türü (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "422": {
            "description": "RFC 9457 Problem Details — olası kodlar: REQUEST_VALIDATION",
            "x-error-codes": [
              "REQUEST_VALIDATION"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "REQUEST_VALIDATION": {
                    "value": {
                      "type": "/errors/request-validation",
                      "title": "İstek doğrulanamadı",
                      "status": 422,
                      "code": "REQUEST_VALIDATION",
                      "detail": "İstek doğrulanamadı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "REQUEST_VALIDATION (alan yolu olmayan hata)": {
                    "value": {
                      "type": "/errors/request-validation",
                      "title": "İstek doğrulanamadı",
                      "status": 422,
                      "code": "REQUEST_VALIDATION",
                      "detail": "İstek gövdesi veya parametreleri doğrulanamadı.",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": [
                        {
                          "field": null,
                          "message": "Field required"
                        },
                        {
                          "field": "top_n",
                          "message": "Input should be greater than or equal to 1"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "500": {
            "description": "RFC 9457 Problem Details — olası kodlar: INTERNAL_ERROR",
            "x-error-codes": [
              "INTERNAL_ERROR"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "/errors/internal-error",
                      "title": "Beklenmeyen hata",
                      "status": 500,
                      "code": "INTERNAL_ERROR",
                      "detail": "Beklenmeyen hata (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "501": {
            "description": "RFC 9457 Problem Details — olası kodlar: NOT_IMPLEMENTED",
            "x-error-codes": [
              "NOT_IMPLEMENTED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "NOT_IMPLEMENTED": {
                    "value": {
                      "type": "/errors/not-implemented",
                      "title": "Uç nokta henüz uygulanmadı",
                      "status": 501,
                      "code": "NOT_IMPLEMENTED",
                      "detail": "Uç nokta henüz uygulanmadı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_upload_api_v1_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_upload_api_v1_uploads_post"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UPLOAD_TOO_LARGE",
          "UPLOAD_INVALID_TYPE",
          "UPLOAD_CORRUPT_OR_ENCRYPTED",
          "SHEET_OR_COLUMN_NOT_FOUND",
          "REQUEST_VALIDATION",
          "INVALID_MODEL",
          "INVALID_PROMPT",
          "COST_LIMIT_EXCEEDED",
          "PROVIDER_AUTH_FAILED",
          "PROVIDER_RATE_LIMITED",
          "PROVIDER_BAD_RESPONSE",
          "PROVIDER_TIMEOUT",
          "JOB_NOT_FOUND",
          "JOB_CONFLICT",
          "NOT_IMPLEMENTED",
          "SERVICE_NOT_READY",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "Dondurulmuş hata sözlüğü ve BE-02 readiness genişletmesi."
      },
      "ErrorItem": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorItem",
        "description": "Alan bazlı doğrulama hatası."
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trace Id"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          },
          "retry_after": {
            "type": "number",
            "minimum": 0.0,
            "title": "Retry After"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "trace_id",
          "errors"
        ],
        "title": "ProblemDetails",
        "description": "ADR-0001 §7 / ADR-0002 #6.\n\n`type`, `title`, `status`, `code`, `detail` ve `trace_id` HER cevapta\nbulunur — frontend'in `problemDetailsSchema`'sı altısını da zorunlu tutar.\n\nBu garantiyi veren `json_schema_serialization_defaults_required` artık\n`ApiModel`'de; burada tekrarlanmıyor. Bayrağın YALNIZCA bu modelde olması,\ndiğer sekiz cevap modelinin openapi.json'da yanlış `required` listesiyle\nyayımlanmasının sebebiydi."
      },
...
      "UploadCreated": {
        "properties": {
          "upload_id": {
            "type": "string",
            "format": "uuid",
            "title": "Upload Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "status"
        ],
        "title": "UploadCreated",
        "description": "POST /api/v1/uploads — 202 Accepted."
      },
...
    },
    "securitySchemes": {
      "OpenRouterKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OpenRouter-Key",
        "description": "BYOK OpenRouter anahtarı. Yalnızca bu header'da taşınır; gövdeye veya sorgu parametresine konmaz, loglarda redakte edilir ve kalıcı olarak saklanmaz."
      }
    }
  }
}